+++ web/CHANGELOG.md
@@ -39,3 +39,24 @@
 ### Compatibility (for it to works well)
 
 -   nothing
+
+## [v0.1.0] 2019-03-31
+
+### Added
+
+-   serviceWorker (sw.js) for offline service
+-   manifest for mobile
+-   not_found page as default when a page is not in cache or internet
+-   temporal logo on the static folder because is required for PWA (in concret for the manifest)
+
+### Changed
+
+-   now the web is a PWA (Progresive Web App)
+
+### Deleted
+
+-   nothing
+
+### Compatibility (for it to works well)
+
+-   nothing